memo: add Distribution info to scans and adjust distribution costs
This commit adds a Distribution physical property to ScanPrivate and
moves locality-optimized search costing adjustment out of computeScanCosts.
SetOp costs with a limit hint are updated. Scans include a distribution
cost as a local scan may not have a Distribute operator.
LargeDistributeCostWithHomeRegion is also added to prefer query plans
with a home region when the `enforce_home_region` session setting is on.

Release note: None
