+++ pkg/sql/opt/xform/coster.go
@@ -163,9 +163,29 @@ const (
 	// stale.
 	largeMaxCardinalityScanCostPenalty = unboundedMaxCardinalityScanCostPenalty / 2
 
+	// DistributeCost is the per-operation cost overhead for Distribute operations
+	// or scans which access remote regions. This is set to a value in the
+	// ballpark of lookup join overhead costs, but should be refined.
+	// TODO(msirek): Measure actual latencies between regions and produce a table
+	//               for determining the maximum latency between the most remote
+	//               region in a distribution and the gateway region.
+	DistributeCost = randIOCostFactor
+
 	// LargeDistributeCost is the cost to use for Distribute operations when a
 	// session mode is set to error out on access of rows from remote regions.
+	// hugeCost cannot be used for this because index hinting used hugeCost to
+	// "force" use of an index. If a LargeDistributeCost of hugeCost were added to
+	// a plan with the forced index, it may cause a plan with different index to
+	// get selected, and error out.
 	LargeDistributeCost = hugeCost / 100
+
+	// LargeDistributeCostWithHomeRegion is the cost to use for Distribute
+	// operations when a session mode is set to error out on access of rows from
+	// remote regions, and the query plan has a home region.
+	// TODO(msirek): Is there a better way of preferring plans that have a home
+	//               region instead of relying on costing, which may not guarantee
+	//               the correct plan is found?
+	LargeDistributeCostWithHomeRegion = LargeDistributeCost / 2
 
 	// preferLookupJoinFactor is a scale factor for the cost of a lookup join when
 	// we have a hint for preferring a lookup join.
@@ -527,7 +547,7 @@ func (c *coster) ComputeCost(candidate memo.RelExpr, required *physical.Required
 
 	case opt.UnionOp, opt.IntersectOp, opt.ExceptOp,
 		opt.UnionAllOp, opt.IntersectAllOp, opt.ExceptAllOp, opt.LocalityOptimizedSearchOp:
-		cost = c.computeSetCost(candidate)
+		cost = c.computeSetCost(candidate, required)
 
 	case opt.GroupByOp, opt.ScalarGroupByOp, opt.DistinctOnOp, opt.EnsureDistinctOnOp,
 		opt.UpsertDistinctOnOp, opt.EnsureUpsertDistinctOnOp:
@@ -678,12 +698,18 @@ func (c *coster) computeDistributeCost(
 		return memo.Cost(0)
 	}
 	if c.evalCtx != nil && c.evalCtx.SessionData().EnforceHomeRegion && c.evalCtx.Planner.IsANSIDML() {
+		if distribute.HasHomeRegion() {
+			// Query plans with a home region are favored over those without one.
+			return LargeDistributeCostWithHomeRegion
+		}
 		return LargeDistributeCost
 	}
 
-	// TODO(rytaft): Compute a real cost here. Currently we just add a tiny cost
-	// as a placeholder.
-	return cpuCostFactor
+	// TODO(rytaft,msirek): Compute a real cost here. Currently we just add a cost
+	//                      which is on par with a lookup (in lookup join) or
+	//                      single-span read costs as a rough estimate of
+	//                      overhead, but actual measurements would be useful.
+	return DistributeCost
 }
 
 func (c *coster) computeScanCost(scan *memo.ScanExpr, required *physical.Required) memo.Cost {
@@ -773,6 +799,42 @@ func (c *coster) computeScanCost(scan *memo.ScanExpr, required *physical.Require
 	}
 
 	cost := baseCost + memo.Cost(rowCount)*(seqIOCostFactor+perRowCost)
+
+	var regionsAccessed physical.Distribution
+	if scan.Distribution.Regions != nil {
+		regionsAccessed = scan.Distribution
+	} else {
+		tabMeta := scan.Memo().Metadata().TableMeta(scan.Table)
+		regionsAccessed.FromIndexScan(c.ctx, c.evalCtx, tabMeta, scan.Index, scan.Constraint)
+	}
+	// Scans that read rows outside of the gateway region incur a distribution
+	// cost.
+	if !regionsAccessed.Any() && !scan.LocalityOptimized {
+		extraCost := memo.Cost(DistributeCost)
+		if c.evalCtx != nil && c.evalCtx.SessionData().EnforceHomeRegion &&
+			c.evalCtx.Planner.IsANSIDML() {
+			if len(regionsAccessed.Regions) == 1 {
+				// Query plans with a home region are favored over those without one.
+				extraCost = LargeDistributeCostWithHomeRegion
+			} else {
+				extraCost = LargeDistributeCost
+			}
+		}
+		if len(regionsAccessed.Regions) > 1 {
+			// Non-multiregion tables may have no regions populated in
+			// regionsAccessed. To avoid potential plan regressions involving
+			// non-multiregion tables, don't add a distribution cost when
+			// `regionsAccessed.Any()` is true because query planning can't be done in
+			// that case to try and avoid the distribution anyway.
+			cost += extraCost
+		} else {
+			var localDist physical.Distribution
+			localDist.FromLocality(c.evalCtx.Locality)
+			if !localDist.Equals(regionsAccessed) {
+				cost += extraCost
+			}
+		}
+	}
 	return cost
 }
 
@@ -937,11 +999,13 @@ func (c *coster) computeLookupJoinCost(
 		provided := distribution.BuildLookupJoinLookupTableDistribution(c.ctx, c.evalCtx, join)
 		if provided.Any() || len(provided.Regions) != 1 {
 			cost += LargeDistributeCost
-		}
-		var localDist physical.Distribution
-		localDist.FromLocality(c.evalCtx.Locality)
-		if !localDist.Equals(provided) {
-			cost += LargeDistributeCost
+		} else {
+			var localDist physical.Distribution
+			localDist.FromLocality(c.evalCtx.Locality)
+			if !localDist.Equals(provided) {
+				// Query plans with a home region are favored over those without one.
+				cost += LargeDistributeCostWithHomeRegion
+			}
 		}
 	}
 	return cost
@@ -1103,11 +1167,13 @@ func (c *coster) computeInvertedJoinCost(
 		provided := distribution.BuildInvertedJoinLookupTableDistribution(c.ctx, c.evalCtx, join)
 		if provided.Any() || len(provided.Regions) != 1 {
 			cost += LargeDistributeCost
-		}
-		var localDist physical.Distribution
-		localDist.FromLocality(c.evalCtx.Locality)
-		if !localDist.Equals(provided) {
-			cost += LargeDistributeCost
+		} else {
+			var localDist physical.Distribution
+			localDist.FromLocality(c.evalCtx.Locality)
+			if !localDist.Equals(provided) {
+				// Query plans with a home region are favored over those without one.
+				cost += LargeDistributeCostWithHomeRegion
+			}
 		}
 	}
 	return cost
@@ -1219,9 +1285,12 @@ func isStreamingSetOperator(relation memo.RelExpr) bool {
 	return false
 }
 
-func (c *coster) computeSetCost(set memo.RelExpr) memo.Cost {
+func (c *coster) computeSetCost(set memo.RelExpr, required *physical.Required) memo.Cost {
 	// Add the CPU cost of emitting the rows.
 	outputRowCount := set.Relational().Statistics().RowCount
+	if outputRowCount != 0 && required.LimitHint != 0 && outputRowCount > required.LimitHint {
+		outputRowCount = required.LimitHint
+	}
 	cost := memo.Cost(outputRowCount) * cpuCostFactor
 
 	// A set operation must process every row from both tables once. UnionAll and
