+++ pkg/sql/opt/xform/groupby_funcs.go
@@ -564,6 +564,7 @@ func (c *CustomFuncs) GenerateLimitedGroupByScans(
 
 		// Scan whatever columns we need which are available from the index.
 		newScanPrivate := *sp
+		newScanPrivate.Distribution.Regions = nil
 		newScanPrivate.Index = index.Ordinal()
 		newScanPrivate.Cols = indexCols.Intersection(sp.Cols)
 		// If the index is not covering, scan the needed index columns plus
